+++ src/lib/ares_mkquery.c
@@ -20,7 +20,7 @@
 #include "ares.h"
 
 int ares_mkquery(const char *name, int dnsclass, int type, unsigned short id,
-                 int rd, unsigned char **buf, int *buflen)
+                 int rd, unsigned char **buf, size_t *buflen)
 {
   return ares_create_query(name, dnsclass, type, id, rd, buf, buflen, 0);
 }
